feat: simplifying card render
