+++ packages/lib/src/components/Card/components/ClickToPay/components/CtPOneTimePasswordInput/CtPOneTimePasswordInput.scss
@@ -6,17 +6,24 @@
     color: #0066FF;
     margin-left: auto;
     cursor: pointer;
+}
 
-    &--disabled {
-        color: $color-gray-darker;
-        cursor: default;
-    }
+.adyen-checkout-ctp__otp-resend-code--disabled {
+    font-size: 13px;
+    font-weight: 400;
+    margin-left: auto;
+    color: $color-gray-darker;
+    cursor: default;
+}
 
-    &-counter {
-        color: black;
-        display: inline-block;
-        text-align: right;
-    }
+.adyen-checkout-ctp__otp-resend-code-counter {
+    font-size: 13px;
+    font-weight: 400;
+    margin-left: auto;
+    cursor: default;
+    color: black;
+    display: inline-block;
+    text-align: right;
 }
 
 .adyen-checkout__field.adyen-checkout__field--otp {
